+++ influunt-api/modules/influunt-simulador/app/simulador/akka/SimuladorAkka.java
@@ -26,21 +26,21 @@ public class SimuladorAkka extends Simulador {
 
 
     @Override
-    public void onEstagioChange(int anel, Long numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos) {
+    public void onEstagioChange(int anel, int numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos) {
         if (timestamp.compareTo(simuladorActor.getPagina()) >= 0) {
             simuladorActor.storeEstagio(anel, timestamp.plus(intervalos.getDuracao()), intervalos);
         }
     }
 
     @Override
-    public void onEstagioEnds(int anel, Long numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos) {
+    public void onEstagioEnds(int anel, int numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos) {
         if (timestamp.compareTo(simuladorActor.getPagina()) >= 0) {
             simuladorActor.storeEstagio(anel, timestamp, intervalos);
         }
     }
 
     @Override
-    public void onCicloEnds(int anel, Long numeroCiclos) {
+    public void onCicloEnds(int anel, int numeroCiclos) {
 
     }
